requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.0
Flask==1.1.1
gcloud==0.18.3
googleapis-common-protos==1.51.0
httplib2==0.16.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.3
jws==0.1.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.2
protobuf==3.11.2
pusher-push-notifications==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.4
PyJWT==1.7.1
pylint==2.1.1
Pyrebase4==4.3.0
python-jwt==2.0.1
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.11.0
urllib3==1.24
Werkzeug==0.16.0
wrapt==1.10.11
